Move JWT Token and CurrentAbility methods into concerns to be mixed into both the ApplicationController and the ProxyController (which subclasses ActiveStorage::BaseController)
