+++ app/controllers/concerns/custom_pod_ability_concern.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+##
+# A concern to add the a custom current_ability
+module CustomPodAbilityConcern
+  def current_ability
+    @current_ability ||= Ability.new(current_user, jwt_token)
+  end
+end
